python_backend/requirements.txt
async-generator==1.10
attrs==21.4.0
certifi==2021.10.8
cffi==1.15.0
click==8.0.3
cryptography==36.0.1
distro==1.7.0
Flask==2.0.3
h11==0.13.0
idna==3.3
iniconfig==1.1.1
itsdangerous==2.1.0
Jinja2==3.0.3
lxml==4.8.0
MarkupSafe==2.1.0
numpy==1.22.2
outcome==1.1.0
packaging==21.3
pandas==1.4.1
pluggy==1.0.0
py==1.11.0
pycparser==2.21
pyOpenSSL==22.0.0
pyparsing==3.0.7
pytest==7.0.1
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
selenium==4.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
tabula-py==2.3.0
tomli==2.0.1
trio==0.19.0
trio-websocket==0.9.2
urllib3==1.26.8
Werkzeug==2.0.3
wsproto==1.0.0
